+++ examples/solidstart-minimal-starter/src/routes/api/trpc/[trpc].ts
@@ -1,6 +1,5 @@
 import { createSolidApiHandler } from '@trpc/server/adapters/solid';
-import { createContext } from '~/server/trpc/context';
-import { appRouter } from '~/server/trpc/router/_app';
+import { appRouter, createContext } from '~/server/trpc';
 
 export const { GET, POST } = createSolidApiHandler({
   router: appRouter,
