--- examples/solidstart-minimal-starter/src/server/trpc/router/_app.ts
@@ -1,8 +1,0 @@
-import { router } from "../utils";
-import example from "./example";
-
-export const appRouter = router({
-  example,
-});
-
-export type IAppRouter = typeof appRouter;
