--- examples/solidstart-minimal-starter/src/server/trpc/router/example.ts
@@ -1,13 +1,0 @@
-import { z } from "zod";
-import { procedure, router } from "../utils";
-
-export default router({
-  hello: procedure.input(z.object({ name: z.string() })).query(({ input }) => {
-    return `Hello ${input.name}`;
-  }),
-  random: procedure
-    .input(z.object({ num: z.number() }))
-    .mutation(({ input }) => {
-      return Math.floor(Math.random() * 100) / input.num;
-    }),
-});
