--- examples/solidstart-minimal-starter/src/server/trpc/utils.ts
@@ -1,7 +1,0 @@
-import { initTRPC } from "@trpc/server";
-import type { IContext } from "./context";
-
-export const t = initTRPC.context<IContext>().create();
-
-export const router = t.router;
-export const procedure = t.procedure;
