+++ examples/solidstart-minimal-starter/src/utils/trpc.ts
@@ -1,14 +1,14 @@
-import { QueryClient } from "@adeora/solid-query";
-import type { IAppRouter } from "~/server/trpc/router/_app";
-import { createTRPCSolid } from "solid-trpc";
-import { httpBatchLink } from "@trpc/client";
+import { QueryClient } from '@adeora/solid-query';
+import { httpBatchLink } from '@trpc/client';
+import { createTRPCSolid } from 'solid-trpc';
+import { AppRouter } from '~/server/trpc';
 
 const getBaseUrl = () => {
-  if (typeof window !== "undefined") return "";
+  if (typeof window !== 'undefined') return '';
   return `http://localhost:${process.env.PORT ?? 5173}`;
 };
 
-export const trpc = createTRPCSolid<IAppRouter>();
+export const trpc = createTRPCSolid<AppRouter>();
 export const client = trpc.createClient({
   links: [
     httpBatchLink({
